refactored to fetch using suspense
